requirements.txt
boto3==1.33.1
colorlog==6.8.2
docker==7.0.0
Flask==3.0.3
holidays==0.46
hyperopt==0.2.7
imbalanced-learn==0.11.0
joblib==1.4.0
lightgbm==4.3.0
mlflow==2.11.3
numba==0.59.1
numpy==1.26.4
pandas==2.2.1
pip-upgrader==1.4.15
pyarrow==15.0.2
python-json-logger==2.0.7
PyYAML==6.0.1
requests==2.31.0
s3fs==2024.3.1
s3transfer==0.8.0
scikit-learn==1.4.0
scipy==1.13.0
seaborn==0.13.2
shap==0.45.0
statsmodels==0.14.1
streamlit==1.33.0
xgboost==2.0.3
tqdm==4.66.5
tsfresh==0.20.2
plotly==5.24.1
watchdog==5.0.3